combining CastAbilityInterface and SlotCastingInterface
